Move clip() implementation from generic to frame/series
