+++ databricks/koalas/frame.py
@@ -26,7 +26,9 @@ import pandas as pd
 from pandas.api.types import is_datetime64_dtype, is_datetime64tz_dtype, is_list_like
 from pyspark import sql as spark
 from pyspark.sql import functions as F, Column
-from pyspark.sql.types import BooleanType, StructField, StructType, to_arrow_type
+from pyspark.sql.types import (BooleanType, ByteType, DecimalType, DoubleType, FloatType,
+                               IntegerType, LongType, ShortType, StructField, StructType,
+                               to_arrow_type)
 from pyspark.sql.utils import AnalysisException
 
 from databricks import koalas as ks  # For running doctests and reference resolution in PyCharm.
@@ -1674,6 +1676,68 @@ defaultdict(<class 'list'>, {'col..., 'col...})]
             self._sdf = sdf
         else:
             return DataFrame(sdf, self._metadata.copy())
+
+    def clip(self, lower: Union[float, int] = None, upper: Union[float, int] = None) \
+            -> 'DataFrame':
+        """
+        Trim values at input threshold(s).
+
+        Assigns values outside boundary to boundary values.
+
+        Parameters
+        ----------
+        lower : float or int, default None
+            Minimum threshold value. All values below this threshold will be set to it.
+        upper : float or int, default None
+            Maximum threshold value. All values above this threshold will be set to it.
+
+        Returns
+        -------
+        DataFrame
+            DataFrame with the values outside the clip boundaries replaced.
+
+        Examples
+        --------
+        >>> ks.DataFrame({'A': [0, 2, 4]}).clip(1, 3)
+           A
+        0  1
+        1  2
+        2  3
+
+        Notes
+        -----
+        One difference between this implementation and pandas is that running
+        pd.DataFrame({'A': ['a', 'b']}).clip(0, 1) will crash with "TypeError: '<=' not supported
+        between instances of 'str' and 'int'" while ks.DataFrame({'A': ['a', 'b']}).clip(0, 1)
+        will output the original DataFrame, simply ignoring the incompatible types.
+        """
+        if is_list_like(lower) or is_list_like(upper):
+            raise ValueError("List-like value are not supported for 'lower' and 'upper' at the " +
+                             "moment")
+
+        if lower is None and upper is None:
+            return self
+
+        sdf = self._sdf
+
+        numeric_types = (DecimalType, DoubleType, FloatType, ByteType, IntegerType, LongType,
+                         ShortType)
+        numeric_columns = [c for c in self.columns
+                           if isinstance(sdf.schema[c].dataType, numeric_types)]
+        nonnumeric_columns = [c for c in self.columns
+                              if not isinstance(sdf.schema[c].dataType, numeric_types)]
+
+        if lower is not None:
+            sdf = sdf.select(*[F.when(F.col(c) < lower, lower).otherwise(F.col(c)).alias(c)
+                               for c in numeric_columns] + nonnumeric_columns)
+        if upper is not None:
+            sdf = sdf.select(*[F.when(F.col(c) > upper, upper).otherwise(F.col(c)).alias(c)
+                               for c in numeric_columns] + nonnumeric_columns)
+
+        # Restore initial column order
+        sdf = sdf.select(list(self.columns))
+
+        return ks.DataFrame(sdf)
 
     def head(self, n=5):
         """
